requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0

# Payment Processing
stripe==6.7.0
modern-treasury==2.12.0
requests==2.31.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7

# GUI Framework
tkinter-modern==1.4.4

# Async Support
asyncio==3.4.3
aiohttp==3.8.6

# Security
cryptography==41.0.4
bcrypt==4.0.1
PyJWT==2.8.0

# Data Processing
pandas==2.1.1
numpy==1.24.4

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-asyncio==0.21.1

# Development
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Monitoring & Logging
prometheus-client==0.17.1

# API Documentation
flask-restx==1.1.0
marshmallow==3.20.1